Rename s/PatternInput/PatternImport
