+++ src/features/pattern_import/PatternImport.scss
@@ -1,4 +1,4 @@
-#pattern-input-comments {
+#pattern-import-comments {
   background-color: rgba(0, 0, 0, 0.05);
   border: 1px solid rgba(0, 0, 0, 0.15);
 }
